chore(ci): add tag resolution for npm releases based on package version; (#6404)
